+++ cpp/cameracontroller.h
