+++ cpp/gpscontroller.cpp
