+++ cpp/gpscontroller.h
