+++ cpp/logger.h
