+++ cpp/satcontroller.cpp
