+++ cpp/satcontroller.h
