+++ cpp/transreceiver.cpp
